refactor: extract helpers from chatui/mod.rs
Move apply_setting, fetch_usage, rebuild_display_messages to
chatui/helpers.rs. Terminal setup/teardown to lifecycle.rs.
run() is now a thinner orchestrator. No behavior changes.
